chore: doc-store rename — document_db / document-db / DocumentDb → doc_store / doc-store / DocStore (#55)
추상화 이름이 backend tech (PostgreSQL) 와 어긋남. "document db" 가 RDBMS / NoSQL
함의 끌고 와 혼란 야기. 실 의미 — "어떤 backend 든 문서 저장" — 를 명확히 표현
하는 "doc store" 로 일괄 rename.

영향 (코드):
- shared/document_db/ → shared/doc_store/
- mcp/document-db/ → mcp/doc-store/
- Python 패키지: document_db_mcp → doc_store_mcp
- 클래스: DocumentDbClient → DocStoreClient
- FastMCP 서버 식별자: "document-db" → "doc-store"

영향 (인프라):
- compose 서비스: document-db-mcp → doc-store-mcp
- 컨테이너: dev-team-document-db-mcp → dev-team-doc-store-mcp
- 이미지: dev-team/document-db-mcp:latest → dev-team/doc-store-mcp:latest
- 환경변수: DOCUMENT_DB_MCP_URL → DOC_STORE_MCP_URL

영향 (문서):
- docs/document-db-schema.md → docs/doc-store-schema.md
- 모든 import 경로 / 참조 자동 갱신

검증:
- 단위 테스트 80/80 통과 (shared 55 / chronicler 8 / mcp/doc-store 17)
- docker compose 회귀 — UG → P → CHR → Doc Store MCP 한 사이클 정상
  - session 1 영속 (UG/P 양쪽 publish, dedup)
  - items 2 영속 (user message + agent reply)

동작 변화: 0 (순수 rename).
