+++ chronicler/src/chronicler/processors/session_end.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 import logging
 from typing import ClassVar
 
-from dev_team_shared.document_db import (
+from dev_team_shared.doc_store import (
     AgentSessionUpdate,
-    DocumentDbClient,
+    DocStoreClient,
 )
 from dev_team_shared.event_bus.events import A2AEvent, SessionEndEvent
 
@@ -21,7 +21,7 @@ class SessionEndProcessor(EventProcessor):
 
     event_type: ClassVar[type[A2AEvent]] = SessionEndEvent
 
-    async def process(self, event: A2AEvent, db: DocumentDbClient) -> None:
+    async def process(self, event: A2AEvent, db: DocStoreClient) -> None:
         assert isinstance(event, SessionEndEvent)
         session = await db.agent_session_find_by_context(event.context_id)
         if session is None:
