+++ mcp/doc-store/CLAUDE.md
