+++ mcp/doc-store/src/doc_store_mcp/__init__.py
