+++ mcp/doc-store/src/doc_store_mcp/config.py
