+++ mcp/doc-store/src/doc_store_mcp/db.py
