+++ mcp/doc-store/src/doc_store_mcp/repositories/__init__.py
@@ -1,0 +1,23 @@
+"""Repository 레이어 — collection 별 CRUD 구현."""
+
+from doc_store_mcp.repositories.agent_item import AgentItemRepository
+from doc_store_mcp.repositories.agent_session import AgentSessionRepository
+from doc_store_mcp.repositories.agent_task import AgentTaskRepository
+from doc_store_mcp.repositories.base import (
+    AbstractRepository,
+    ListFilter,
+    PostgresRepositoryBase,
+)
+from doc_store_mcp.repositories.issue import IssueRepository
+from doc_store_mcp.repositories.wiki_page import WikiPageRepository
+
+__all__ = [
+    "AbstractRepository",
+    "AgentItemRepository",
+    "AgentSessionRepository",
+    "AgentTaskRepository",
+    "IssueRepository",
+    "ListFilter",
+    "PostgresRepositoryBase",
+    "WikiPageRepository",
+]
