+++ mcp/doc-store/src/doc_store_mcp/repositories/agent_session.py
@@ -7,8 +7,8 @@ from uuid import UUID
 
 import asyncpg
 
-from document_db_mcp.repositories.base import PostgresRepositoryBase
-from dev_team_shared.document_db.schemas.agent_session import (
+from doc_store_mcp.repositories.base import PostgresRepositoryBase
+from dev_team_shared.doc_store.schemas.agent_session import (
     AgentSessionCreate,
     AgentSessionRead,
     AgentSessionUpdate,
