+++ mcp/doc-store/src/doc_store_mcp/repositories/base.py
