+++ mcp/doc-store/src/doc_store_mcp/tools/agent_item.py
@@ -5,12 +5,12 @@ from __future__ import annotations
 from typing import Any
 from uuid import UUID
 
-from dev_team_shared.document_db.schemas.agent_item import AgentItemCreate, AgentItemRead
-from dev_team_shared.document_db.tool_names import AgentItemTools
+from dev_team_shared.doc_store.schemas.agent_item import AgentItemCreate, AgentItemRead
+from dev_team_shared.doc_store.tool_names import AgentItemTools
 from mcp.server.fastmcp import Context
 
-from document_db_mcp.mcp_instance import AppContext, mcp
-from document_db_mcp.repositories.base import ListFilter
+from doc_store_mcp.mcp_instance import AppContext, mcp
+from doc_store_mcp.repositories.base import ListFilter
 
 
 def _ctx(ctx: Context) -> AppContext:
