--- mcp/document-db/src/document_db_mcp/repositories/__init__.py
@@ -1,23 +1,0 @@
-"""Repository 레이어 — collection 별 CRUD 구현."""
-
-from document_db_mcp.repositories.agent_item import AgentItemRepository
-from document_db_mcp.repositories.agent_session import AgentSessionRepository
-from document_db_mcp.repositories.agent_task import AgentTaskRepository
-from document_db_mcp.repositories.base import (
-    AbstractRepository,
-    ListFilter,
-    PostgresRepositoryBase,
-)
-from document_db_mcp.repositories.issue import IssueRepository
-from document_db_mcp.repositories.wiki_page import WikiPageRepository
-
-__all__ = [
-    "AbstractRepository",
-    "AgentItemRepository",
-    "AgentSessionRepository",
-    "AgentTaskRepository",
-    "IssueRepository",
-    "ListFilter",
-    "PostgresRepositoryBase",
-    "WikiPageRepository",
-]
