+++ shared/src/dev_team_shared/doc_store/__init__.py
@@ -1,12 +1,12 @@
 """Document DB MCP SDK — schemas + 도구명 상수 + typed client.
 
-server (mcp/document-db) 와 client (chronicler / 향후 librarian) 모두 본 모듈을
+server (mcp/doc-store) 와 client (chronicler / 향후 librarian) 모두 본 모듈을
 공유 contract 로 import. wire-level 디테일 (도구명 / dict args / JSON parse) 은
-DocumentDbClient 안에 격리되어 외부로 새지 않음.
+DocStoreClient 안에 격리되어 외부로 새지 않음.
 """
 
-from dev_team_shared.document_db.client import DocumentDbClient
-from dev_team_shared.document_db.schemas import (
+from dev_team_shared.doc_store.client import DocStoreClient
+from dev_team_shared.doc_store.schemas import (
     AgentItemCreate,
     AgentItemRead,
     AgentSessionCreate,
@@ -22,7 +22,7 @@ from dev_team_shared.document_db.schemas import (
     WikiPageRead,
     WikiPageUpdate,
 )
-from dev_team_shared.document_db.tool_names import (
+from dev_team_shared.doc_store.tool_names import (
     AgentItemTools,
     AgentSessionTools,
     AgentTaskTools,
@@ -42,7 +42,7 @@ __all__ = [
     "AgentTaskRead",
     "AgentTaskTools",
     "AgentTaskUpdate",
-    "DocumentDbClient",
+    "DocStoreClient",
     "IssueCreate",
     "IssueRead",
     "IssueTools",
