+++ shared/src/dev_team_shared/doc_store/schemas/wiki_page.py
